* Fixes for changes in DirectEve v1.330.4904.26383

+++ Questor/Behaviors/DirectionalScannerBehavior.cs
@@ -286,7 +286,7 @@ namespace Questor.Behaviors
 
                 case DirectionalScannerBehaviorState.Traveler:
                     Cache.Instance.OpenWrecks = false;
-                    List<long> destination = Cache.Instance.DirectEve.Navigation.GetDestinationPath();
+                    List<int> destination = Cache.Instance.DirectEve.Navigation.GetDestinationPath();
                     if (destination == null || destination.Count == 0)
                     {
                         // happens if autopilot is not set and this QuestorState is chosen manually
